backend/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
openai>=1.0.0
python-dotenv>=1.0.0